Make action_for_key a weak symbol
This fixes the compile error in and re-applies
c17b8a599e3ec3b0a327bcd66082541f2517ab30.

Its purpose is to help with some cleanup in the Planck Keyboard's
fork. It overrides this function to add some custom code before
running a copy-and-pasted version of the default action_for_key.